pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "device-activity-telegram-bot"
version = "1.0.0"
description = "Telegram bot for device activity monitoring and remote shutdown"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "python-telegram-bot>=21.6",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
]
# Platform-specific optional dependencies are handled via requirements-dev.txt

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["halt", "send"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"